Consolidate event handling into single module
